requirements.txt
tensorflow>=2.10.0
numpy>=1.21.0
matplotlib>=3.5.0
Pillow>=9.0.0
imgaug>=0.4.0
Augmentor>=0.2.10
scikit-learn>=1.1.0
seaborn>=0.11.0
jupyter>=1.0.0
ipykernel>=6.0.0